\name{party.votes.xml}
\alias{party.votes.xml}
\title{Sample XML data for party votes from the NY Times.}
\description{Sample XML data for party votes from the NY Times.}
\details{Sample XML data for party votes from the NY Times.}
\docType{data}
\author{Shane Conway \email{shane.conway@gmail.com}}
\seealso{\code{\link{getNYTCongress}}}
\keyword{data}
