\name{roll.call.xml}
\alias{roll.call.xml}
\title{Sample XML data for a roll call vote from the NY Times.}
\description{Sample XML data for a roll call vote from the NY Times.}
\details{Sample XML data for a roll call vote from the NY Times.}
\docType{data}
\author{Shane Conway \email{shane.conway@gmail.com}}
\seealso{\code{\link{getNYTCongress}}}
\keyword{data}
